% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osci-reml-files.r
\name{osci.reml.files}
\alias{osci.reml.files}
\title{Run OSCA REML analysis (filename interface)}
\usage{
osci.reml.files(myorm, pheno, out)
}
\arguments{
\item{pheno}{a phenotype file with FID, IID columns and a phenotype column}

\item{out}{the base name of the output files}

\item{myord}{the base name of the orm files}
}
\value{
A text file with the same name as the input file, a bod file, and an orm file.
}
\description{
Performs REML analysis in OSCA
for a single omic 
given filenames as inputs.
}
