% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osci-reml-multi-files.r
\name{osci.reml.multi.files}
\alias{osci.reml.multi.files}
\title{Run OSCA multi-omic REML analysis (filename interface)}
\usage{
osci.reml.multi.files(file.list, pheno, out)
}
\arguments{
\item{file.list}{a file with the name of each orm file to be included in the analysis on a a separate line.}

\item{pheno}{a phenotype file with FID, IID columns and a phenotype column}

\item{out}{the base name of the output files}
}
\value{
A text file with the same name as the input file, a bod file, and an orm file.
}
\description{
Performs a multi-omic REML analysis in OSCA
given filenames as inputs.
}
