% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osci-reml-multi.r
\name{osci.reml.multi}
\alias{osci.reml.multi}
\title{Run OSCA multi-omic REML analysis}
\usage{
osci.reml.multi(x, ...)
}
\arguments{
\item{x}{Either an ORM list filename or phenotype vector or matrix.
If the former, then osci.reml.multi.files is called, otherwise osci.reml.matrices is called.}

\item{...}{Additional arguments to be passed to osci.reml.multi.files or osci.reml.matrices.}
}
\value{
Returned outputs depend on the value of x, either osci.reml.multi.files or osci.reml.matrices outputs.
}
\description{
Performs REML analysis in OSCA
for a single omic.
}
