% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osci-reml.r
\name{osci.reml}
\alias{osci.reml}
\title{Run OSCA REML analysis}
\usage{
osci.reml(x, ...)
}
\arguments{
\item{x}{Either an ORM filename or a phenotype vector or matrix.
If the former, then osci.reml.files is called, otherwise osci.reml.matrices is called.}

\item{...}{Additional arguments to be passed to osci.reml.files or osci.reml.matrices.}
}
\value{
Returned outputs depend on the value of x, either osci.reml.files or osci.reml.matrices outputs.
}
\description{
Performs REML analysis in OSCA
for a single omic.
}
