% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.r
\name{osci.use_container}
\alias{osci.use_container}
\title{Choose whether to use the OSCA container or not}
\usage{
osci.use_container(use_container = TRUE)
}
\arguments{
\item{use_container}{If TRUE, OSCA will be run inside the container, otherwise in the current environment.}
}
\description{
This function allows the user to choose whether to run OSCA
in the current environment or inside the osci container.
}
