% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osci-write-orm.r
\name{osci.write.orm}
\alias{osci.write.orm}
\title{Write OSCA bod and orm file}
\usage{
osci.write.orm(df, filename)
}
\arguments{
\item{df}{A data frame}

\item{filename}{a file name with a .txt extension for saving the data frame in osca format.}
}
\value{
A text file with the same name as the input file, a bod file, and an orm file.
}
\description{
This function writes a data frame to a text file, then uses osca to make bod and orm files.
}
