% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osci-reml-matrices.r
\name{osci.reml.matrices}
\alias{osci.reml.matrices}
\title{Run OSCA REML analysis (R variable interface)}
\usage{
osci.reml.matrices(var, ..., tmp_dir = NULL)
}
\arguments{
\item{var}{A numeric phenotype vector or matrix with samples in rows and phenotypes in columns.}

\item{...}{Named omic matrices, one matrix per omic, with samples in columns and features in rows.}

\item{tmp_dir}{Optional temporary directory for OSCA input and output files. If NULL,
a temporary directory will be created using tempfile().}
}
\value{
A list with three elements
('stat'--variance estimates, 'se'--standard errors, and 'test'--test statistics) that collate
results across phenotypes and omics.
}
\description{
Performs single or multi-omic REML analysis in OSCA
for multiple phenotypes in parallel.
}
