% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osci.reml.multi.r
\name{osci.reml.multi}
\alias{osci.reml.multi}
\title{Run OSCA multi-orm reml analysis}
\usage{
osci.reml.multi(file.list, pheno, out)
}
\arguments{
\item{file.list}{a file with the name of each orm file to be included in the analysis on a a separate line.}

\item{pheno}{a phenotype file with FID, IID columns and a phenotype column}

\item{out}{the base name of the output files}
}
\value{
A text file with the same name as the input file, a bod file, and an orm file.
}
\description{
This function runs the reml analysis in OSCA.
}
