% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osci-reml.r
\name{osci.reml}
\alias{osci.reml}
\title{Run OSCA reml analysis}
\usage{
osci.reml(myorm, pheno, out)
}
\arguments{
\item{pheno}{a phenotype file with FID, IID columns and a phenotype column}

\item{out}{the base name of the output files}

\item{myord}{the base name of the orm files}
}
\value{
A text file with the same name as the input file, a bod file, and an orm file.
}
\description{
This function runs the reml analysis in OSCA.
}
